Main python script to start the loops.
This script sets in motion the Metaformers stack with selected models. It will ask you for a topic, how many iterations and finally the intervals in which Dolphin3 intervene and asks a meta-question. It will also create a folder in the directory you ran the script from named "runs" inside there will be another timestamped directory and in that directory your logs of the convo with a resume of each turn at the end of each turn done by Dolphin3 so it is faster to read.